application/requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
mysql-connector-python==8.0.33
python-dotenv==0.21.0
gunicorn==21.2.0
flask-cors==4.0.0
